AWS/CT/requirements.txt
boto3==1.26.101
scikit-learn==1.2.2
pandas==1.4.4
numpy==1.22.4
Pillow==8.4.0